updated code

+++ trainer/craft/utils/inference_boxes.py
@@ -234,7 +234,7 @@ def load_icdar2015_gt(dataFolder, isTraing=False):
             word = ",".join(word)
             box_points = np.array(box_points, np.int32).reshape(4, 2)
             cv2.polylines(
-                image, [np.array(box_points).astype(np.int)], True, (0, 0, 255), 1
+                image, [np.array(box_points).astype(np.int32)], True, (0, 0, 255), 1
             )
             box_info_dict["points"] = box_points
             box_info_dict["text"] = word
